[Fix]: seporate hadlepasswordcheck function and early return
